Add option to not call start_new_game in set_fen_position.
This could be useful if going from a position back to its parent position, since the TT could still hold relevant data. The default behaviour for the function is to still call start_new_game (and thus reset the TT), in order to preserve backward compatibility with anyone currently using this function.